+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/CodeGeneration/CodeGenerationOptions.cs
@@ -6,11 +6,7 @@ using System.Runtime.Serialization;
 using Microsoft.CodeAnalysis.AddImport;
 using Microsoft.CodeAnalysis.CodeStyle;
 using Microsoft.CodeAnalysis.Diagnostics.Analyzers.NamingStyles;
-using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Options;
-//#if !CODE_STYLE
-//using Microsoft.CodeAnalysis.Host;
-//#endif
 
 namespace Microsoft.CodeAnalysis.CodeGeneration;
 
@@ -34,24 +30,4 @@ internal record CodeGenerationOptions
     {
         NamingStyle = options.GetOption(NamingStyleOptions.NamingPreferences, language);
     }
-
-    public static CodeGenerationOptions GetDefault(LanguageServices languageServices)
-        => languageServices.GetRequiredService<ICodeGenerationService>().DefaultOptions;
-}
-
-[DataContract]
-internal readonly record struct CodeAndImportGenerationOptions
-{
-    [DataMember]
-    public required CodeGenerationOptions GenerationOptions { get; init; }
-
-    [DataMember]
-    public required AddImportPlacementOptions AddImportOptions { get; init; }
-
-    internal static CodeAndImportGenerationOptions GetDefault(LanguageServices languageServices)
-        => new()
-        {
-            GenerationOptions = CodeGenerationOptions.GetDefault(languageServices),
-            AddImportOptions = AddImportPlacementOptions.Default
-        };
 }
