+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/CodeGeneration/CodeGenerationOptionsProviders.cs
@@ -6,11 +6,9 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.AddImport;
 using Microsoft.CodeAnalysis.CodeCleanup;
-using Microsoft.CodeAnalysis.Formatting;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.Shared.Extensions;
-using Microsoft.CodeAnalysis.Simplification;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CodeGeneration;
@@ -49,10 +47,6 @@ internal static class CodeGenerationOptionsProviders
         return service.GetInfo(context, options, document.Project.ParseOptions);
     }
 
-    public static CodeCleanupOptions GetDefault(LanguageServices languageServices)
-        => new()
-        {
-            FormattingOptions = SyntaxFormattingOptions.GetDefault(languageServices),
-            SimplifierOptions = SimplifierOptions.GetDefault(languageServices)
-        };
+    public static CodeGenerationOptions GetDefault(LanguageServices languageServices)
+        => languageServices.GetRequiredService<ICodeGenerationService>().DefaultOptions;
 }
